Abstracción del Logger fuera del bot.py
